requirements.txt
asgiref==3.7.2
colorama==0.4.6
Django==4.2.7
django-phonenumber-field==7.2.0
djangorestframework==3.14.0
iniconfig==2.0.0
packaging==23.2
phonenumbers==8.13.25
pluggy==1.3.0
pytz==2023.3.post1
sqlparse==0.4.4
tinydb==4.8.0
tzdata==2023.3